removed usages of Error class
