+++ packages/Verifiers/LeagueDictVerifier.py
@@ -34,3 +34,17 @@ class LeagueDictVerifier:
                     if len(teamIdsThatHavePlayed) != len(teamIdsSet):
                         return True
         return False
+
+    @staticmethod
+    def duplicateTeamNames(years: dict) -> bool:
+        """
+        Returns a boolean of whether the given years dict has any years duplicate team names.
+        """
+        for year in years:
+            teamNames = []
+            for team in years[year]["teams"]:
+                teamNames.append(team["teamName"].lower().strip())
+            teamNamesSet = set(teamNames)
+            if len(teamNames) != len(teamNamesSet):
+                return True
+        return False
