--- test/test_helpers/test_Error.py
@@ -1,10 +1,0 @@
-import unittest
-
-from helpers.Error import Error
-
-
-class TestError(unittest.TestCase):
-
-    def test_errorMessage(self):
-        error = Error("test")
-        self.assertEqual("test", error.errorMessage())
